% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{mstep}
\alias{mstep}
\title{M-Step of the EM Algorithm}
\usage{
mstep(x, Tnk)
}
\arguments{
\item{x}{An \eqn{n \times d}{n x d} matrix of observations.}

\item{Tnk}{An \eqn{n \times k}{n x k} matrix of posterior responsibilities from the E-step,
where \eqn{T_{i,j}} is the probability that observation \eqn{i} belongs to component \eqn{j}.}
}
\value{
A list containing updated parameter estimates:
\item{Tau}{Updated mixture proportions, a numeric vector of length \eqn{k}}
\item{Mu}{List of updated mean vectors, each of length \eqn{d}}
\item{Sigma}{List of updated covariance matrices, each \eqn{d \times d}{d x d}}
}
\description{
Maximizes the expected complete log-likelihood to update parameter estimates.
This is the maximization step of the EM algorithm.
}
\details{
The updates are computed as:
\deqn{\tau_j^{(t+1)} = \frac{1}{n} \sum_{i=1}^{n} T_{i,j}}
\deqn{\mu_j^{(t+1)} = \frac{\sum_{i=1}^{n} T_{i,j} x_i}{\sum_{i=1}^{n} T_{i,j}}}
\deqn{\Sigma_j^{(t+1)} = \frac{\sum_{i=1}^{n} T_{i,j} (x_i - \mu_j^{(t+1)})(x_i - \mu_j^{(t+1)})'}{\sum_{i=1}^{n} T_{i,j}}}
}
\keyword{internal}
